GeneralsX @feature Fix 4 additional enum-float conversion warnings
- HelicopterSlowDeathUpdate.cpp: 2 warnings (division by LOGICFRAMES_PER_SECOND)
- AIUpdate.cpp: 2 warnings (division by LOGICFRAMES_PER_SECOND in formula)
- ChinookAIUpdate.cpp: 1 warning (multiplication by LOGICFRAMES_PER_SECOND)

Total fixed: 32/71 (45% complete)
Remaining: 39 warnings in 30+ files

Strategy: Continue systematic fixing of 1-2 warning files in batches

+++ GeneralsMD/Code/GameEngine/Source/GameLogic/Object/Update/HelicopterSlowDeathUpdate.cpp
@@ -318,7 +318,7 @@ UpdateSleepTime HelicopterSlowDeathBehavior::update( void )
 			{
 
 				// we're going towards the max self spin, increase it
-				m_selfSpin += modData->m_selfSpinUpdateAmount / LOGICFRAMES_PER_SECOND;
+				m_selfSpin += modData->m_selfSpinUpdateAmount / static_cast<float>(LOGICFRAMES_PER_SECOND);
 				if( m_selfSpin > modData->m_maxSelfSpin )
 				{
 
@@ -332,7 +332,7 @@ UpdateSleepTime HelicopterSlowDeathBehavior::update( void )
 			{
 
 				// we're going towards the min self spin, decrease it
-				m_selfSpin -= modData->m_selfSpinUpdateAmount / LOGICFRAMES_PER_SECOND;
+				m_selfSpin -= modData->m_selfSpinUpdateAmount / static_cast<float>(LOGICFRAMES_PER_SECOND);
 				if( m_selfSpin < modData->m_minSelfSpin )
 				{
